pass config instead of instance, publish as ESM
